Working into cli tool. Testing the waters.
