+++ bin/cli.rb
@@ -1,0 +1,7 @@
+#!/usr/bin/env ruby
+
+#-wU
+
+require 'cfnmason'
+
+CfnMason::HammerOfTheGods.start( ARGV )
