--- lib/cfn/mason.rb
@@ -1,9 +1,0 @@
-require "cfn/mason/version"
-require "cfn/mason/hello"
-require "cfn/mason/cli"
-
-module Cfn 
-  module Mason
-    # Your code goes here...
-  end
-end
