--- lib/cfn/mason/hello.rb
@@ -1,9 +1,0 @@
-module Cfn
-  module Mason
-    class << self
-      def hi
-        puts "Hello Cruel World."
-      end
-    end
-  end
-end
