--- lib/cfn/mason/version.rb
@@ -1,5 +1,0 @@
-module Cfn
-  module Mason
-    VERSION = "0.1.1"
-  end
-end
