+++ lib/cfnmason.rb
@@ -1,0 +1,6 @@
+require "cfnmason/version"
+require "cfnmason/hello"
+require "cfnmason/cli"
+
+module CfnMason
+end
