+++ lib/cfnmason/hello.rb
@@ -1,0 +1,7 @@
+module CfnMason
+  class << self
+    def hi
+      puts "Hello Cruel World."
+    end
+  end
+end
