+++ lib/cfnmason/version.rb
@@ -1,0 +1,3 @@
+module CfnMason
+  VERSION = "0.1.1"
+end
